Fix for package cl is deprecated
Currently when using key-combo, in the Messages buffer is a note that package cl is deprecated.

With these changes, the `cl-lib` package is used instead of `cl`.  I byte-compiled the code and no longer had no warnings about functions in the cl package.  (There are other warnings about free variables, but that's for a different patch.)